+++ HISTORY.rst
@@ -15,6 +15,8 @@ PlatformIO 3.0
 
 * Simplify configuration for `PIO Unit Testing <http://docs.platformio.org/page/plus/unit-testing.html>`__: separate main program from a test build process, drop
   requirement for ``#ifdef UNIT_TEST`` guard
+* Override any option from board manifest in `Project Configuration File "platformio.ini" <http://docs.platformio.org/page/projectconf/section_env_board.html#more-options>`__
+  (`issue #1612 <https://github.com/platformio/platformio-core/issues/1612>`_)
 * Configure a custom path to SVD file using `debug_svd_path <http://docs.platformio.org/page/projectconf/section_env_debug.html#debug-svd-path>`__
   option
 * Custom project `description <http://docs.platformio.org/en/latest/projectconf/section_platformio.html#description>`_